[Refactor] [2/N] Move tool parsers into the vLLM main directory
