+++ vllm/tool_parsers/llama4_pythonic_tool_parser.py
@@ -18,10 +18,10 @@ from vllm.entrypoints.openai.protocol import (
     FunctionCall,
     ToolCall,
 )
-from vllm.entrypoints.openai.tool_parsers.abstract_tool_parser import (
+from vllm.logger import init_logger
+from vllm.tool_parsers.abstract_tool_parser import (
     ToolParser,
 )
-from vllm.logger import init_logger
 
 logger = init_logger(__name__)
