+++ vllm/tool_parsers/openai_tool_parser.py
@@ -12,10 +12,10 @@ from vllm.entrypoints.openai.protocol import (
     FunctionCall,
     ToolCall,
 )
-from vllm.entrypoints.openai.tool_parsers.abstract_tool_parser import (
+from vllm.logger import init_logger
+from vllm.tool_parsers.abstract_tool_parser import (
     ToolParser,
 )
-from vllm.logger import init_logger
 
 if TYPE_CHECKING:
     from vllm.tokenizers import TokenizerLike
